improved app structure
