+++ src/app.js
@@ -1,131 +1,16 @@
 'use strict';
 
 const express = require('express');
-const app = express();
-
 const bodyParser = require('body-parser');
-const logger = require('./logger');
+const healthRouter = require('./routes/health');
+const ridesRouter = require('./routes/rides');
+
 const jsonParser = bodyParser.json();
 
-const DEFAULT_OFFSET = 0;
-const DEFAULT_LIMIT = 50;
+const app = express();
 
-module.exports = (db) => {
-  app.get('/health', (req, res) => res.send('Healthy'));
+app.use(jsonParser);
+app.use(healthRouter);
+app.use(ridesRouter);
 
-  app.post('/rides', jsonParser, async (req, res) => {
-    try {
-      const startLatitude = Number(req.body.start_lat);
-      const startLongitude = Number(req.body.start_long);
-      const endLatitude = Number(req.body.end_lat);
-      const endLongitude = Number(req.body.end_long);
-      const riderName = req.body.rider_name;
-      const driverName = req.body.driver_name;
-      const driverVehicle = req.body.driver_vehicle;
-
-      if (startLatitude < -90 || startLatitude > 90 || startLongitude < -180 || startLongitude > 180) {
-        return res.send({
-          error_code: 'VALIDATION_ERROR',
-          message: 'Start latitude and longitude must be between -90 - 90 and -180 to 180 degrees respectively',
-        });
-      }
-
-      if (endLatitude < -90 || endLatitude > 90 || endLongitude < -180 || endLongitude > 180) {
-        return res.send({
-          error_code: 'VALIDATION_ERROR',
-          message: 'End latitude and longitude must be between -90 - 90 and -180 to 180 degrees respectively',
-        });
-      }
-
-      if (typeof riderName !== 'string' || riderName.length < 1) {
-        return res.send({
-          error_code: 'VALIDATION_ERROR',
-          message: 'Rider name must be a non empty string',
-        });
-      }
-
-      if (typeof driverName !== 'string' || driverName.length < 1) {
-        return res.send({
-          error_code: 'VALIDATION_ERROR',
-          message: 'Rider name must be a non empty string',
-        });
-      }
-
-      if (typeof driverVehicle !== 'string' || driverVehicle.length < 1) {
-        return res.send({
-          error_code: 'VALIDATION_ERROR',
-          message: 'Rider name must be a non empty string',
-        });
-      }
-
-      const values = [
-        req.body.start_lat,
-        req.body.start_long,
-        req.body.end_lat,
-        req.body.end_long,
-        req.body.rider_name,
-        req.body.driver_name,
-        req.body.driver_vehicle,
-      ];
-
-      const result = await db.run(
-          `INSERT INTO Rides(startLat, startLong, endLat, endLong, riderName, driverName, driverVehicle) 
-        VALUES (?, ?, ?, ?, ?, ?, ?)`,
-          values);
-      const rows = await db.all('SELECT * FROM Rides WHERE rideID = ?', result.lastID);
-
-      res.send(rows);
-    } catch (err) {
-      logger.error(err);
-      return res.send({
-        error_code: 'SERVER_ERROR',
-        message: 'Unknown error',
-      });
-    }
-  });
-
-  app.get('/rides', async (req, res) => {
-    try {
-      const offset = req.query.offset || DEFAULT_OFFSET;
-      const limit = req.query.limit || DEFAULT_LIMIT;
-
-      const rows = await db.all('SELECT * FROM Rides ORDER BY rideID LIMIT ? OFFSET ?', limit, offset);
-      if (rows.length === 0) {
-        return res.send({
-          error_code: 'RIDES_NOT_FOUND_ERROR',
-          message: 'Could not find any rides',
-        });
-      }
-
-      res.send(rows);
-    } catch (err) {
-      logger.error(err);
-      return res.send({
-        error_code: 'SERVER_ERROR',
-        message: 'Unknown error',
-      });
-    }
-  });
-
-  app.get('/rides/:id', async (req, res) => {
-    try {
-      const rows = await db.all(`SELECT * FROM Rides WHERE rideID='${req.params.id}'`);
-      if (rows.length === 0) {
-        return res.send({
-          error_code: 'RIDES_NOT_FOUND_ERROR',
-          message: 'Could not find any rides',
-        });
-      }
-
-      res.send(rows);
-    } catch (err) {
-      logger.error(err);
-      return res.send({
-        error_code: 'SERVER_ERROR',
-        message: 'Unknown error',
-      });
-    }
-  });
-
-  return app;
-};
+module.exports = app;
